+++ react_front/package.json
@@ -12,7 +12,8 @@
   },
   "dependencies": {
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.0"
   },
   "devDependencies": {
     "@types/react": "^18.3.3",
@@ -26,7 +27,9 @@
     "eslint-plugin-react-refresh": "^0.4.7",
     "jest": "^29.7.0",
     "postcss": "^8.4.41",
+    "styled-components": "^6.1.12",
     "tailwindcss": "^3.4.7",
+    "twin.macro": "^3.4.1",
     "typescript": "^5.2.2",
     "vite": "^5.3.4"
   }
