+++ react_front/src/components/navbar.tsx
@@ -1,0 +1,7 @@
+
+
+export function Navbar() {
+    return (
+        <div></div>
+    );
+}
